requirements.txt
fastapi>=0.116.0
fastapi[standard]
uvicorn>=0.35.0
langchain>=0.3.0
langchain-openai>=0.3.0
azure-cosmos>=4.9.0
azure-search-documents>=11.5.0
azure-identity>=1.25.0
langchain-community>=0.3.2
pydantic-settings>=2.10.0
PyPDF2>=3.0.0
python-docx>=1.2.0
